+++ CSharpRegexTools4Npp/PluginInfrastructure/Scintilla_iface.cs
@@ -38,36 +38,36 @@ namespace CSharpRegexTools4Npp.PluginInfrastructure
     public struct ScNotification
     {
         public ScNotificationHeader Header;
-        private readonly int position;               /* SCN_STYLENEEDED, SCN_DOUBLECLICK, SCN_MODIFIED, SCN_MARGINCLICK, SCN_NEEDSHOWN, SCN_DWELLSTART, SCN_DWELLEND, SCN_CALLTIPCLICK, SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, SCN_INDICATORCLICK, SCN_INDICATORRELEASE, SCN_USERLISTSELECTION, SCN_AUTOCSELECTION */
-        public int character;               /* SCN_CHARADDED, SCN_KEY, SCN_AUTOCCOMPLETE, SCN_AUTOCSELECTION, SCN_USERLISTSELECTION */
-        public int Mmodifiers;              /* SCN_KEY, SCN_DOUBLECLICK, SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, SCN_INDICATORCLICK, SCN_INDICATORRELEASE */
-        public int ModificationType;        /* SCN_MODIFIED - modification types are name "SC_MOD_*" */
+        private readonly uint position;               /* SCN_STYLENEEDED, SCN_DOUBLECLICK, SCN_MODIFIED, SCN_MARGINCLICK, SCN_NEEDSHOWN, SCN_DWELLSTART, SCN_DWELLEND, SCN_CALLTIPCLICK, SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, SCN_INDICATORCLICK, SCN_INDICATORRELEASE, SCN_USERLISTSELECTION, SCN_AUTOCSELECTION */
+        public uint character;               /* SCN_CHARADDED, SCN_KEY, SCN_AUTOCCOMPLETE, SCN_AUTOCSELECTION, SCN_USERLISTSELECTION */
+        public uint Mmodifiers;              /* SCN_KEY, SCN_DOUBLECLICK, SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, SCN_INDICATORCLICK, SCN_INDICATORRELEASE */
+        public uint ModificationType;        /* SCN_MODIFIED - modification types are name "SC_MOD_*" */
         public IntPtr TextPointer;          /* SCN_MODIFIED, SCN_USERLISTSELECTION, SCN_AUTOCSELECTION, SCN_URIDROPPED */
-        public int Length;                  /* SCN_MODIFIED */
-        public int LinesAdded;              /* SCN_MODIFIED */
-        public int Message;                 /* SCN_MACRORECORD */
+        public uint Length;                  /* SCN_MODIFIED */
+        public uint LinesAdded;              /* SCN_MODIFIED */
+        public uint Message;                 /* SCN_MACRORECORD */
         public IntPtr wParam;               /* SCN_MACRORECORD */
         public IntPtr lParam;               /* SCN_MACRORECORD */
 
         /// <summary>
         /// 0-based index
         /// </summary>
-        public int LineNumber;           /* SCN_MODIFIED */
-        public int FoldLevelNow;         /* SCN_MODIFIED */
-        public int FoldLevelPrev;        /* SCN_MODIFIED */
-        public int Margin;               /* SCN_MARGINCLICK */
-        public int ListType;             /* SCN_USERLISTSELECTION */
-        public int X;                    /* SCN_DWELLSTART, SCN_DWELLEND */
-        public int Y;                    /* SCN_DWELLSTART, SCN_DWELLEND */
-        public int Token;                /* SCN_MODIFIED with SC_MOD_CONTAINER */
-        public int AnnotationLinesAdded; /* SC_MOD_CHANGEANNOTATION */
-        public int Updated;              /* SCN_UPDATEUI */
-        public int ListCompletionMethod; /* SCN_AUTOCSELECTION, SCN_AUTOCCOMPLETED, SCN_USERLISTSELECTION */
+        public uint LineNumber;           /* SCN_MODIFIED */
+        public uint FoldLevelNow;         /* SCN_MODIFIED */
+        public uint FoldLevelPrev;        /* SCN_MODIFIED */
+        public uint Margin;               /* SCN_MARGINCLICK */
+        public uint ListType;             /* SCN_USERLISTSELECTION */
+        public uint X;                    /* SCN_DWELLSTART, SCN_DWELLEND */
+        public uint Y;                    /* SCN_DWELLSTART, SCN_DWELLEND */
+        public uint Token;                /* SCN_MODIFIED with SC_MOD_CONTAINER */
+        public uint AnnotationLinesAdded; /* SC_MOD_CHANGEANNOTATION */
+        public uint Updated;              /* SCN_UPDATEUI */
+        public uint ListCompletionMethod; /* SCN_AUTOCSELECTION, SCN_AUTOCCOMPLETED, SCN_USERLISTSELECTION */
 
         /// <summary>
         /// SCN_STYLENEEDED, SCN_DOUBLECLICK, SCN_MODIFIED, SCN_MARGINCLICK, SCN_NEEDSHOWN, SCN_DWELLSTART, SCN_DWELLEND, SCN_CALLTIPCLICK, SCN_HOTSPOTCLICK, SCN_HOTSPOTDOUBLECLICK, SCN_HOTSPOTRELEASECLICK, SCN_INDICATORCLICK, SCN_INDICATORRELEASE, SCN_USERLISTSELECTION, SCN_AUTOCSELECTION
         /// </summary>
-        public Position Position { get { return new Position(position); } }
+        public Position Position { get { return new Position((int)position); } }
 
         /// <summary>
         /// Character of the notification - eg keydown
